Created new endpoints and updated others

+++ backend/src/resources/campaign/campaign.service.ts
@@ -1,14 +1,12 @@
-import { Request, Response } from "express";
-import { Campaign, PrismaClient } from "@prisma/client";
+import { PrismaClient } from "@prisma/client";
 import { CampaignDto, CreateCampaignDto, UpdateCampaignDto } from "./campaign.types";
 
 const prisma = new PrismaClient();
 
 export const createCampaign = async (
   campaign: CreateCampaignDto,
-  req: Request,
+  uid: string,
 ): Promise<CampaignDto> => {
-  const userId = req.session.uid as string;
   return await prisma.campaign.create({
     select: {
       id: true,
@@ -25,33 +23,49 @@ export const createCampaign = async (
     },
     data: {
       ...campaign,
-      userId: userId,
+      userId: uid,
     },
   });
 };
 
-export const listCampaigns = async (skip?: number, take?: number): Promise<CampaignDto[]> => {
-  return await prisma.campaign.findMany({
-    select: {
-      id: true,
-      goal: true,
-      deadline: true,
-      title: true,
-      description: true,
-      preview: true,
-      category: true,
-      imageUrl: true,
-      userId: true,
-      createdAt: true,
-      updatedAt: true,
-    },
-    skip,
-    take,
-  });
+export const listCampaigns = async (
+  searchTerm: string,
+  skip?: number,
+  take?: number,
+): Promise<CampaignDto[]> => {
+  if (searchTerm) {
+    return await prisma.campaign.findMany({
+      where: {
+        title: {
+          contains: searchTerm,
+        },
+      },
+      skip: skip,
+      take: take,
+    });
+  } else {
+    return await prisma.campaign.findMany({
+      select: {
+        id: true,
+        goal: true,
+        deadline: true,
+        title: true,
+        description: true,
+        preview: true,
+        category: true,
+        imageUrl: true,
+        userId: true,
+        createdAt: true,
+        updatedAt: true,
+      },
+      skip,
+      take,
+    });
+  }
 };
 
 export const listUserCampaigns = async (
-  req: Request,
+  uid: string,
   skip?: number,
   take?: number,
 ): Promise<CampaignDto[]> => {
@@ -69,7 +83,7 @@ export const listUserCampaigns = async (
       createdAt: true,
       updatedAt: true,
     },
-    where: { userId: req.session.uid },
+    where: { userId: uid },
     skip,
     take,
   });
@@ -97,56 +111,14 @@ export const readCampaign = async (id: string): Promise<CampaignDto | null> => {
 export const updateCampaign = async (
   id: string,
   updatedCampaign: UpdateCampaignDto,
-  req: Request,
-): Promise<Campaign | null> => {
-  const campaign = await prisma.campaign.findUnique({
-    select: { userId: true },
-    where: { id },
-  });
-
-  if (!campaign) {
-    throw new Error("Campanha não encontrada");
-  }
-
-  if (campaign.userId !== req.session.uid) {
-    throw new Error("Usuário não autorizado a atualizar esta campanha");
-  }
-
+  uid: string,
+): Promise<CampaignDto | null> => {
   return await prisma.campaign.update({
-    where: { id },
+    where: { id: id, userId: uid },
     data: updatedCampaign,
   });
 };
 
-export const deleteCampaign = async (id: string, req: Request): Promise<Campaign> => {
-  const campaign = await prisma.campaign.findUnique({
-    select: { userId: true },
-    where: { id },
-  });
-
-  if (!campaign) {
-    throw new Error("Campanha não encontrada");
-  }
-
-  if (campaign.userId !== req.session.uid) {
-    throw new Error("Usuário não autorizado a remover esta campanha");
-  }
-  return await prisma.campaign.delete({ where: { id } });
-};
-
-export const searchCampaigns = async (
-  req: Request,
-  searchTerm: string,
-  skip?: number,
-  take?: number,
-) => {
-  return await prisma.campaign.findMany({
-    where: {
-      title: {
-        contains: searchTerm,
-      },
-    },
-    skip: skip,
-    take: take,
-  });
+export const deleteCampaign = async (id: string, uid: string): Promise<CampaignDto> => {
+  return await prisma.campaign.delete({ where: { id: id, userId: uid } });
 };
